archive/2026-01-25-old-flask-backend/requirements.txt
# P2P Data Products Flask Backend - Requirements
# ================================================
# Python dependencies for Flask-based REST API

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# HANA Database Driver
hdbcli==2.19.21

# Environment Variables
python-dotenv==1.0.0

# Development Dependencies
pytest==7.4.3
pytest-flask==1.3.0
